+++ providers/coingecko/coinlist.go
@@ -1,0 +1,7 @@
+package coingecko
+
+type CoinList []struct {
+	ID     string `json:"id"`
+	Symbol string `json:"symbol"`
+	Name   string `json:"name"`
+}