+++ providers/coingecko/prices.go
@@ -1,0 +1,3 @@
+package coingecko
+
+type Prices map[string]map[string]float64