+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 ### Added
 
 1. Configuration via environment file.
+1. New provider Coingecko.
 
 ## [0.2.0] - 2021-05-08